pyproject.toml
[project]
name = "genai-prototype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
]
